+++ lib/glimmer/swt/scrolled_composite_proxy.rb
@@ -1,21 +1,16 @@
 require 'glimmer/swt/widget_proxy'
+require 'glimmer/swt/swt_proxy'
 
 module Glimmer
   module SWT
-    class ScrolledCompositeProxy < Glimmer::SWT::WidgetProxy
-      def initialize(underscored_widget_name, parent, args)
-        unless args.first.is_a?(Numeric)
-          args.unshift(:h_scroll)
-          args.unshift(:v_scroll)
-        end
-        super
-        swt_widget.expand_horizontal = true
-        swt_widget.expand_vertical = true
-      end
-      
+    class ScrolledCompositeProxy < Glimmer::SWT::WidgetProxy    
       def post_initialize_child(child)
         swt_widget.content = child.swt_widget        
       end
+      
+      def post_add_content
+        swt_widget.set_min_size(swt_widget.computeSize(SWTProxy[:default], SWTProxy[:default]))
+      end      
     end
   end
 end